feat: remove iamctl
BREAKING CHANGE: Use the standard IAM command from
                 https://github.com/einride/aip-cli-go
