+++ .github/dependabot.yml
@@ -12,11 +12,6 @@ updates:
       interval: daily
 
   - package-ecosystem: gomod
-    directory: /cmd/iamctl
-    schedule:
-      interval: daily
-
-  - package-ecosystem: gomod
     directory: .sage
     schedule:
       interval: weekly
