--- cmd/iamctl/internal/examplecmd/exampleclientcmd/shipments_list.go
@@ -1,72 +1,0 @@
-package exampleclientcmd
-
-import (
-	"context"
-	"log"
-
-	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-	"go.einride.tech/iam/cmd/iamctl/internal/connection"
-	iamexamplev1 "go.einride.tech/iam/proto/gen/einride/iam/example/v1"
-	"google.golang.org/protobuf/encoding/protojson"
-)
-
-// nolint: gochecknoglobals
-var listShipmentsCommand = &cobra.Command{
-	Use:   "list-shipments",
-	Short: "List shipments",
-	RunE: func(cmd *cobra.Command, args []string) error {
-		viperCfg := viper.New()
-		if err := viperCfg.BindPFlags(cmd.Flags()); err != nil {
-			return err
-		}
-		if err := viperCfg.BindPFlags(cmd.PersistentFlags()); err != nil {
-			return err
-		}
-		var flags listShipmentsFlags
-		if err := viperCfg.Unmarshal(&flags); err != nil {
-			return err
-		}
-		conn, err := flags.Connect(cmd.Context())
-		if err != nil {
-			return err
-		}
-		defer func() {
-			_ = conn.Close()
-		}()
-		client := iamexamplev1.NewFreightServiceClient(conn)
-		return runListShipmentCommand(cmd.Context(), client, &flags)
-	},
-}
-
-type listShipmentsFlags struct {
-	connection.Flags `mapstructure:",squash"`
-	Parent           string `mapstructure:"parent"`
-	PageSize         int32  `mapstructure:"page-size"`
-	PageToken        string `mapstructure:"page-token"`
-}
-
-// nolint: gochecknoinits
-func init() {
-	listShipmentsCommand.Flags().String("parent", "", "parent shipper")
-	listShipmentsCommand.Flags().Int32("page-size", 0, "page size")
-	listShipmentsCommand.Flags().String("page-token", "", "page token")
-	_ = listShipmentsCommand.MarkFlagRequired("parent")
-}
-
-func runListShipmentCommand(
-	ctx context.Context,
-	client iamexamplev1.FreightServiceClient,
-	flags *listShipmentsFlags,
-) error {
-	response, err := client.ListShipments(ctx, &iamexamplev1.ListShipmentsRequest{
-		Parent:    flags.Parent,
-		PageSize:  flags.PageSize,
-		PageToken: flags.PageToken,
-	})
-	if err != nil {
-		return err
-	}
-	log.Println(protojson.Format(response))
-	return nil
-}
